Init replace NFS with Volume
