+++ modal/_utils/mount_utils.py
@@ -6,10 +6,9 @@ from typing import Dict, List, Mapping, Sequence, Tuple, Union
 
 from ..cloud_bucket_mount import _CloudBucketMount
 from ..exception import InvalidError
-from ..network_file_system import _NetworkFileSystem
 from ..volume import _Volume
 
-T = typing.TypeVar("T", bound=Union["_Volume", "_NetworkFileSystem", "_CloudBucketMount"])
+T = typing.TypeVar("T", bound=Union["_Volume", "_CloudBucketMount"])
 
 
 def validate_mount_points(
@@ -37,7 +36,7 @@ def validate_mount_points(
 
 def validate_volumes(
     volumes: Mapping[Union[str, PurePosixPath], Union["_Volume", "_CloudBucketMount"]],
-) -> Sequence[Tuple[str, Union["_Volume", "_NetworkFileSystem", "_CloudBucketMount"]]]:
+) -> Sequence[Tuple[str, Union["_Volume", "_CloudBucketMount"]]]:
     if not isinstance(volumes, dict):
         raise InvalidError("volumes must be a dict where the keys are mount paths")
 
@@ -46,7 +45,7 @@ def validate_volumes(
     # but the same CloudBucketMount object can be used in more than one location.
     volume_to_paths: Dict["_Volume", List[str]] = {}
     for path, volume in validated_volumes:
-        if not isinstance(volume, (_Volume, _NetworkFileSystem, _CloudBucketMount)):
+        if not isinstance(volume, (_Volume, _CloudBucketMount)):
             raise InvalidError(f"Object of type {type(volume)} mounted at '{path}' is not useable as a volume.")
         elif isinstance(volume, _Volume):
             volume_to_paths.setdefault(volume, []).append(path)
